excelRead

+++ LoginForm/LoginForm/DataSet/ExtendedRange.cs
@@ -45,6 +45,6 @@ namespace LoginForm.DataSet
         public Nullable<decimal> DiscountedPrice4 { get; set; }
         public Nullable<decimal> DiscountedPrice5 { get; set; }
         public string ManufacturerCode { get; set; }
-        public Nullable<int> ExtendedRangeWeight { get; set; }
+        public Nullable<decimal> ExtendedRangeWeight { get; set; }
     }
 }
